+++ NEWS.md
@@ -1,3 +1,7 @@
+# rhosa development
+
+* New function `biperiodogram()` for calculating the biperiodogram of real-valued time series.
+
 # rhosa 0.2.0
 
 * `cross_bicoherence()` now follow an alternative definition of (cross-)bicoherence due to [Kim and Powers (1979)](https://doi.org/10.1109/TPS.1979.4317207).